requirements.txt
aiohttp
discord.py
python-dotenv
Daug
